# Conflicts:
#	assets/data/simulation.ipynb

+++ src/sumo_gym/typing.py
@@ -17,6 +17,6 @@ SpaceType = Space
 ActionsType = npt.NDArray[int]
 RewardsType = npt.NDArray[float]
 
-FMPElectricVehiclesType = npt.NDArray[Tuple[int]]
-FMPChargingStationType = npt.NDArray[Tuple[int]]
-FMPDemandsType = npt.NDArray[Tuple[int]]
+FMPElectricVehiclesType = npt.NDArray[Tuple[float]]     # vehicle index, speed, indicator, capacity
+FMPChargingStationType = npt.NDArray[Tuple[int]]        # vertex index, indicator, speed
+FMPDemandsType = npt.NDArray[Tuple[int]]                # departure, destination
